subclass the proxy socket
